requirements.txt
# API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# UI
streamlit==1.30.0

# Simulation
simpy==4.1.1

# Data & Charts
matplotlib==3.8.2
pandas==2.2.0
numpy==1.26.3

# Config
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-multipart==0.0.6
python-dateutil==2.8.2
